+++ ProductRegistration/Model/UserDto/SigninDto.cs
@@ -3,7 +3,7 @@ using Infrastructure.Entity.User;
 
 namespace ProductRegistration.Model.UserDto;
 
-public class SigninDto : BaseDto<SigninDto, User, int>
+public class SigninDto : BaseDto<SigninDto, User, long>
 {
     public string UsreName { get; set; }
     public string Password { get; set; }
