+++ ProductRegistration/Model/UserDto/SignupDto.cs
@@ -3,7 +3,7 @@ using Infrastructure.Entity.User;
 
 namespace ProductRegistration.Model.UserDto;
 
-public class SignupDto : BaseDto<SignupDto, User, int>
+public class SignupDto : BaseDto<SignupDto, User, long>
 {
     public string Name { get; set; }
     public string UsreName { get; set; }
